fix: test build cluster + enable tls works
